Moved files into correct sample folder
