--- Samples/constants.py
@@ -1,132 +1,0 @@
-ATTR_MAPPING = {
-    'AS': 'networkSourceAsn',
-    'email-dst': 'emailRecipient',
-    'email-src-display-name': 'emailSenderName',
-    'email-subject': 'emailSubject',
-    'email-x-mailer': 'emailXMailer',
-    'filename': 'fileName',
-    'malware-type': 'malwareFamilyNames',
-    'mutex': 'fileMutexName',
-    'port': 'networkPort',
-    'published': 'isActive',
-    'size-in-bytes': 'fileSize',
-    'url': 'url',
-    'user-agent': 'userAgent',
-    'uuid': 'externalId'
-}
-
-MISP_HASH_TYPES = frozenset([
-    "filename|authentihash",
-    "filename|impfuzzy",
-    "filename|imphash",
-    "filename|md5",
-    "filename|pehash",
-    "filename|sha1",
-    "filename|sha224",
-    "filename|sha256",
-    "filename|sha384",
-    "filename|sha512",
-    "filename|sha512/224",
-    "filename|sha512/256",
-    "filename|ssdeep",
-    "filename|tlsh",
-    "authentihash",
-    "impfuzzy",
-    "imphash",
-    "md5",
-    "pehash",
-    "sha1",
-    "sha224",
-    "sha256",
-    "sha384",
-    "sha512",
-    "sha512/224",
-    "sha512/256",
-    "ssdeep",
-    "tlsh",
-])
-
-MISP_SPECIAL_CASE_TYPES = frozenset([
-    *MISP_HASH_TYPES,
-    'ip-dst',
-    'ip-src',
-    'domain|ip',
-    'email-src',
-    'ip-dst|port',
-    'ip-src|port'
-])
-
-MISP_ACTIONABLE_TYPES = frozenset([
-    *ATTR_MAPPING.keys(),
-    *MISP_SPECIAL_CASE_TYPES
-])
-
-# api class
-REQUIRED_API_METADATA = frozenset([
-    "type", 
-    "spec_version", 
-    "id", 
-    "created", 
-    "modified", 
-    "pattern", 
-    "pattern_type", 
-    "valid_from",
-])
-
-OPTIONAL_API_METADATA = frozenset([
-    "created_by_ref", 
-    "revoked", 
-    "labels", 
-    "confidence", 
-    "lang", 
-    "external_references", 
-    "object_marking_refs", 
-    "granular_markings", 
-    "extensions", 
-    "name", 
-    "description", 
-    "indicator_types", 
-    "pattern_version", 
-    "valid_until", 
-    "kill_chain_phases",
-])
-
-# keep as is for now. 
-API_OBSERVABLES = frozenset([
-    "emailEncoding",
-    "emailLanguage",
-    "emailRecipient",
-    "emailSenderAddress",
-    "emailSenderName",
-    "emailSourceDomain",
-    "emailSourceIPAddress",
-    "emailSubject",
-    "emailXMailer",
-    "fileCompileDateTime",
-    "fileCreationDateTime",
-    "fileHashType",
-    "fileHashValue",
-    "fileMutexName",
-    "fileName",
-    "filePacker",
-    "filePath",
-    "fileSize",
-    "fileType",
-    "domainName",
-    "networkIPv4",
-    "networkIPv6",
-    "networkPort",
-    "networkDestinationAsn",
-    "networkDestinationCidrBlock",
-    "networkDestinationIPv4",
-    "networkDestinationIPv6",
-    "networkDestinationPort",
-    "networkProtocol",
-    "networkSourceAsn",
-    "networkSourceCidrBlock",
-    "networkSourceIPv4",
-    "networkSourceIPv6",
-    "networkSourcePort",
-    "url",
-    "userAgent",
-])
